Use long long for uint64_t/int64_t
	* gcc.target/i386/pr45685.c (uint64_t): Replace long with long
	long.
	(int64_t): Likewise.


git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@221255 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/testsuite/gcc.target/i386/pr45685.c
@@ -1,8 +1,8 @@
 /* { dg-do compile { target { ! { ia32 } } } } */
 /* { dg-options "-O3" } */
 
-typedef unsigned long int uint64_t;
-typedef long int int64_t;
+typedef unsigned long long int uint64_t;
+typedef long long int int64_t;
 int summation_helper_1(int64_t* products, uint64_t count)
 {
 	int s = 0;
